JDF-374 Fix to ensure bookings view is refreshed on delete
This required a server-side change to return HTTP 204 No Content responses instead of HTTP 200 OK. This is primarily due to a change in jQuery 1.9 that treats empty 200 OK responses as errors which cascaded to Backbone and eventually resulted in no updates to the view. The server-side change now results in a 204 response preventing this hidden failure.