JDF-389 Updated Git patches for Forge 1.3.2
This patch also updates a few issues in the TicketMonster application that are not fixed in the Forge scaffolding or cannot be fixed in the Forge scaffolding.
Code in PerformanceBean.java has been enhanced to allow correct deletion of Performances. Code in ShowBean.java and show/create.xhtml has been enhanced to allow correct selection of Venue sections when a different venue is chosen during creation/updation of Shows. RichFaces calendar widgets for datetime values have the Apply button enabled via this patch.