JDF-407 Replaced Errai+GWT based Monitor with HTML5 and Backbone.
This implements a poller based bot and metrics view. As a result,
Errai and GWT are no longer used in the project.

The tutorial now demonstrates how to create a HTML5 and JavaScript
based dashboard.

Deleted Errai related files that were no longer used.
Refactored some classes to comply with a better naming convention.

Updated the architecture diagram.

Corrected the link to the Monitor in the Administration site.