JDF-495 Replaced the JDF Admin UI with a AngularJS based one.
Committing the generated scaffold and REST resources into the
codebase. This also resolves JDF-118. Regenerated the patches
and rewrote the Forge script to use the new commands.

Updated the tutorial to use the AngularJS scaffolding.

Added better labels for displaying Shows and Performances. Better
labels are now displayed for Performances in select and
multiselect widgets.

Added overrides for the default Bootstrap style for buttons.

Added fix to ensure that Performances are deleted correctly.
This required additional code to be added to the generated REST
resource for the Performance entity. Resolves WFK2-59.

Added fix to ensure that Sections specific to a Venue or Show
are displayed when applicable. This prevents a repeat of an
issue similar to WFK2-57.

Improved the datetime handling support. Added a custom AngularJS
directive to support datetime type.

The admin section now uses a custom AngularJS directive that uses
Modernizr to determine whether to use a HTML5 datetime-local
widget with a custom parser and formatter, or a text input based
widget with a different parser and formatter.

Added customizations to display MediaItems for Events and Venues.