JDF-552 Corrected error handling during booking for mobile site.
Human readable error messages are parsed and displayed when they
are available in the HTTP response. Ticket allocations are still
validated on the server-side since they can fail.

Additionally, JS validation is added in the Create Booking screen
to ensure that only valid email addresses and valid ticket
quantities enable the checkout button. Field-level messages are
displayed when the validation fails.