JDF-335 Converted Cordova projects to JBDS Hybrid Mobile project
The individual XCode and Android ADT projects have been replaced
with the JBDS Hybrid Mobile project structure,

A specialized landing page that builds on top of the TicketMonster
index.html is now used. This page is used to setup the mobile
environment for the Cordova app.

Updated the tutorial to use the JBDS tooling.