JDF-711 Override exception mapping for RESTEasy ReaderException.
Created a JAX-RS ExceptionMapper for the RESTEasy ReaderException.
This ensures that the default behavior of the RESTEasy exception mapper
to return HTML responses is overridden.

JSON responses are now returned if the client accepts them.

Also cleaned up the POM and removed direct dependency on Jackson Core.
The RESTEeasy Jackson provider brings this dependency as a transitive one.